+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/KafkaExporter.java
@@ -9,15 +9,11 @@ import io.fabric8.kubernetes.api.model.ContainerBuilder;
 import io.fabric8.kubernetes.api.model.ContainerPort;
 import io.fabric8.kubernetes.api.model.EnvVar;
 import io.fabric8.kubernetes.api.model.HasMetadata;
-import io.fabric8.kubernetes.api.model.IntOrString;
 import io.fabric8.kubernetes.api.model.LocalObjectReference;
 import io.fabric8.kubernetes.api.model.Secret;
 import io.fabric8.kubernetes.api.model.SecurityContext;
 import io.fabric8.kubernetes.api.model.Volume;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
-import io.fabric8.kubernetes.api.model.apps.DeploymentStrategy;
-import io.fabric8.kubernetes.api.model.apps.DeploymentStrategyBuilder;
-import io.fabric8.kubernetes.api.model.apps.RollingUpdateDeploymentBuilder;
 import io.strimzi.api.kafka.model.ContainerEnvVar;
 import io.strimzi.api.kafka.model.Kafka;
 import io.strimzi.api.kafka.model.KafkaClusterSpec;
@@ -149,12 +145,13 @@ public class KafkaExporter extends AbstractModel {
                     kafkaExporter.templateDeploymentAnnotations = template.getDeployment().getMetadata().getAnnotations();
                 }
 
-                if (template.getContainer() != null && template.getContainer().getEnv() != null) {
-                    kafkaExporter.templateContainerEnvVars = template.getContainer().getEnv();
-                }
-
-                if (template.getContainer() != null && template.getContainer().getSecurityContext() != null) {
-                    kafkaExporter.templateContainerSecurityContext = template.getContainer().getSecurityContext();
+                if (template.getContainer() != null) {
+                    if (template.getContainer().getEnv() != null) {
+                        kafkaExporter.templateContainerEnvVars = template.getContainer().getEnv();
+                    }
+                    if (template.getContainer().getSecurityContext() != null) {
+                        kafkaExporter.templateContainerSecurityContext = template.getContainer().getSecurityContext();
+                    }
                 }
 
                 if (template.getServiceAccount() != null && template.getServiceAccount().getMetadata() != null) {
@@ -162,6 +159,7 @@ public class KafkaExporter extends AbstractModel {
                     kafkaExporter.templateServiceAccountAnnotations = template.getServiceAccount().getMetadata().getAnnotations();
                 }
 
+                ModelUtils.parseDeploymentTemplate(kafkaExporter, template.getDeployment());
                 ModelUtils.parsePodTemplate(kafkaExporter, template.getPod());
                 kafkaExporter.templatePodLabels = Util.mergeLabelsOrAnnotations(kafkaExporter.templatePodLabels, DEFAULT_POD_LABELS);
             }
@@ -182,16 +180,8 @@ public class KafkaExporter extends AbstractModel {
     }
 
     public Deployment generateDeployment(boolean isOpenShift, ImagePullPolicy imagePullPolicy, List<LocalObjectReference> imagePullSecrets) {
-        DeploymentStrategy updateStrategy = new DeploymentStrategyBuilder()
-                .withType("RollingUpdate")
-                .withRollingUpdate(new RollingUpdateDeploymentBuilder()
-                        .withMaxSurge(new IntOrString(1))
-                        .withMaxUnavailable(new IntOrString(0))
-                        .build())
-                .build();
-
         return createDeployment(
-                updateStrategy,
+                getDeploymentStrategy(),
                 Collections.emptyMap(),
                 Collections.emptyMap(),
                 getMergedAffinity(),